requirements.txt
Flask==3.1.2
python-pptx==1.0.2
requests==2.32.5
